Move facility-schedule-user routes and fix file naming of publicAppointmentApi
