+++ src/Utils/request/api.tsx
@@ -14,7 +14,6 @@ import {
   ResourceRequest,
   UpdateResourceRequest,
 } from "@/types/resourceRequest/resourceRequest";
-import { UserReadMinimal } from "@/types/user/user";
 
 /**
  * A fake function that returns an empty object casted to type T
@@ -118,16 +117,6 @@ const routes = {
     method: "PUT",
     TRes: Type<{ message: string }>(),
     TBody: Type<UpdatePasswordForm>(),
-  },
-
-  getScheduleAbleFacilityUser: {
-    path: "/api/v1/facility/{facility_id}/schedulable_users/{user_id}/",
-    TRes: Type<UserReadMinimal>(),
-  },
-
-  getScheduleAbleFacilityUsers: {
-    path: "/api/v1/facility/{facility_id}/schedulable_users/",
-    TRes: Type<PaginatedResponse<UserReadMinimal>>(),
   },
 
   // Request
