+++ src/types/facility/publicFacilityApi.ts
@@ -1,5 +1,6 @@
 import { HttpMethod, Type } from "@/Utils/request/api";
 import { PaginatedResponse } from "@/Utils/request/types";
+import { UserReadMinimal } from "@/types/user/user";
 
 import { FacilityPublicRead } from "./facility";
 
@@ -13,5 +14,15 @@ export default {
     path: "/api/v1/getallfacilities/{id}/",
     method: HttpMethod.GET,
     TRes: Type<FacilityPublicRead>(),
+  },
+  listFacilitySchedulableUsers: {
+    path: "/api/v1/facility/{facilityId}/schedulable_users/",
+    method: HttpMethod.GET,
+    TRes: Type<PaginatedResponse<UserReadMinimal>>(),
+  },
+  getFacilitySchedulableUser: {
+    path: "/api/v1/facility/{facilityId}/schedulable_users/{userId}/",
+    method: HttpMethod.GET,
+    TRes: Type<UserReadMinimal>(),
   },
 } as const;
